+++ RELEASE_NOTES.md
@@ -5,7 +5,7 @@
 This update refactors the internal page source model parsing. This will likely not affect you directly, however, if you have written custom code that interacts with any class relating to the PageParser contract, you'll want to take a closer look at the changes.
 
 ### Added
-- for new features.
+- Added a new static shorthand to quickly parse Markdown files into MarkdownDocuments (`MarkdownFileParser::parse()`)
 
 ### Changed
 - All source model parsing is now handled by the new SourceFileParser action